src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
langchain-community==0.0.38
PyMuPDF==1.25.1
google-generativeai==0.8.5
qdrant-client==1.10.1
supabase==2.10.0
